Major bug fixes

+++ main.py
@@ -86,39 +86,44 @@ async def start(bot, update):
         reply_markup=START_BUTTON
     )
 
-@bughunter0.on_message(filters.command(["test"])) # PdfToText NoT Working!!
+@bughunter0.on_message(filters.document | (filters.document & filters.forwarded)) 
+async def document(bot, message):
+  chat_id=int(message.chat.id)
+  bot.send_message(text="Now Use /pdf2txt to Convert it to Text file \n Use /info to Get Information about the PDF file",chat_id=chat_id,reply_markup=ForceReply)
+  
+@bughunter0.on_message(filters.command(["pdf2txt"])) # PdfToText 
 async def pdf_to_text(bot, message):
       try :
-           txt =await message.reply_text("Validating Pdf ")
-           pdf_path = DOWNLOAD_LOCATION + f"{message.chat.id}.pdf" #pdfFileObject
-           await txt.edit("Downloading.....")
-           
-           await message.reply_to_message.download(pdf_path)  
-           await txt.edit("Downloaded File")
-           pdf = open(pdf_path,'rb')
-           pdf_reader = PyPDF2.PdfFileReader(pdf) #pdfReaderObject
-           await txt.edit("Getting Number of Pages....")
-           num_of_pages = pdf_reader.getNumPages()
-           await txt.edit(f"Found {num_of_pages} Page")
-           page_no = pdf_reader.getPage(0) # pageObject
-           await txt.edit("Extracting Text from PDF...")
-           page_content = """ """ # EmptyString   
-           with open(f'{message.chat.id}.txt', 'a+') as text_path:   
-               for page in range (0,num_of_pages):
-                    file_write = open(f'{message.chat.id}.txt','a+') 
-                    page_no = pdf_reader.getPage(page) # Iteration of page number
-                    page_content = page_no.extractText()
-                    file_write.write(f"page number - {page}")
-                    file_write.write(f" {page_content} ")   
-                    file_write.write(f"© BugHunterBots")
-                 #  await message.reply_text(f"**Page Number  :  {page}  **\n\n  ` {page_content} `\n     @BugHunterBots\n\n") # Use this Line of code to get Pdf Text as Messages
+           if message.reply_to_message:
+                pdf_path = DOWNLOAD_LOCATION + f"{message.chat.id}.pdf" #pdfFileObject
+                txt = await message.reply_text("Downloading.....")     
+                await message.reply_to_message.download(pdf_path)  
+                await txt.edit("Downloaded File")
+                pdf = open(pdf_path,'rb')
+                pdf_reader = PyPDF2.PdfFileReader(pdf) #pdfReaderObject
+                await txt.edit("Getting Number of Pages....")
+                num_of_pages = pdf_reader.getNumPages()
+                await txt.edit(f"Found {num_of_pages} Page")
+                page_no = pdf_reader.getPage(0) # pageObject
+                await txt.edit("Extracting Text from PDF...")
+                page_content = """ """ # EmptyString   
+                with open(f'{message.chat.id}.txt', 'a+') as text_path:   
+                  for page in range (0,num_of_pages):
+                      file_write = open(f'{message.chat.id}.txt','a+') 
+                      page_no = pdf_reader.getPage(page) # Iteration of page number
+                      page_content = page_no.extractText()
+                      file_write.write(f"page number - {page}")
+                      file_write.write(f" {page_content} ")   
+                      file_write.write(f"© BugHunterBots")
+                   #  await message.reply_text(f"**Page Number  :  {page}  **\n\n  ` {page_content} `\n     @BugHunterBots\n\n") # Use this Line of code to get Pdf Text as Messages
                         
-           with open(f'{message.chat.id}.txt', 'a+') as text_path:  
-                    await message.reply_document(f"{message.chat.id}.txt",caption="©@BugHunterBots")      
+                with open(f'{message.chat.id}.txt', 'a+') as text_path:  
+                      await message.reply_document(f"{message.chat.id}.txt",caption="©@BugHunterBots")      
          
-           os.remove(pdf_path)
-           os.remove(f"{message.chat.id}.txt")  
-          
+                os.remove(pdf_path)
+                os.remove(f"{message.chat.id}.txt")  
+           else :
+                await message.reply("Please Reply to PDF file")
       except Exception as error :
            await txt.delete()
            await message.reply_text(f"{error}")
@@ -128,19 +133,20 @@ async def pdf_to_text(bot, message):
 @bughunter0.on_message(filters.command(["info"]))
 async def info(bot, message):
      try:
-         txt = await message.reply_text("Validating Pdf ")  
-         pdf_path = DOWNLOAD_LOCATION + f"{message.chat.id}.pdf" #pdfFileObject
-         await txt.edit("Downloading.....")
-         await message.reply_to_message.download(pdf_path)  
-         await txt.edit("Downloaded File")
-         pdf = open(pdf_path,'rb')
-         pdf_reader = PyPDF2.PdfFileReader(pdf) #pdfReaderObject
-         await txt.edit("Getting Number of Pages....")
-         num_of_pages = pdf_reader.getNumPages()
-         await txt.edit(f"Found {num_of_pages} Page")
-         await txt.edit("Getting PDF info..")
-         info = pdf_reader.getDocumentInfo()
-         await txt.edit(f"""
+         if message.reply_to_message:
+              txt = await message.reply_text("Validating Pdf ")  
+              pdf_path = DOWNLOAD_LOCATION + f"{message.chat.id}.pdf" #pdfFileObject
+              await txt.edit("Downloading.....")
+              await message.reply_to_message.download(pdf_path)  
+              await txt.edit("Downloaded File")
+              pdf = open(pdf_path,'rb')
+              pdf_reader = PyPDF2.PdfFileReader(pdf) #pdfReaderObject
+              await txt.edit("Getting Number of Pages....")
+              num_of_pages = pdf_reader.getNumPages()
+              await txt.edit(f"Found {num_of_pages} Page")
+              await txt.edit("Getting PDF info..")
+              info = pdf_reader.getDocumentInfo()
+              await txt.edit(f"""
 **Author :** `{info.author}`
 **Creator :** `{info.creator}`
 **Producer :** `{info.producer}`
@@ -148,11 +154,13 @@ async def info(bot, message):
 **Title :** `{info.title}`
 **Pages :** `{num_of_pages}`""")
 
-         os.remove(pdf_path)
+              os.remove(pdf_path)
+         else:
+             await message.reply_text("Please Reply to a Pdf File")
      except Exception as error :
          await message.reply_text(f"Oops , {error}")
 
-@bughunter0.on_message(filters.command(["merge"]))
+@bughunter0.on_message(filters.command(["merge"])) # Under Maintenance
 async def merge(bot, message):
       try:
          txt = await message.reply_text("Validating Pdf ")
